+++ NEWS.md
@@ -7,6 +7,7 @@
   - Methods for getting covers and pages as PIL Images.
   - Lazy metadata and cover pulling. Removes `metadata` and `cover` options.
   - closefd option leaves archive open for multiple method calls. Close manually with close().
+  - .cbt Tarfile comic archive support.
 
 ## v0.5.0